modules-engine/requirements.txt
# Modules Engine - Master Primitives
# Core dependencies for the mini-app spawning system

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Document Intelligence (Lightweight)
PyMuPDF>=1.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0


# Machine Learning
scikit-learn>=1.3.0
# monai>=1.3.0  # Medical Imaging (Future/Separate App)

# Vector Store / RAG
langchain>=0.1.0
langchain-community>=0.0.10
pgvector>=0.2.0
psycopg2-binary>=2.9.0

# Report Generation
reportlab>=4.0.0
Pillow>=10.0.0
qrcode>=7.4.0
python-barcode>=0.15.0
plotly>=5.18.0

# Notifications
requests>=2.31.0
# africastalking>=1.2.0  # SMS for Kenya

# Validation
pydantic>=2.5.0


# Marimo (reactive notebooks)
marimo>=0.3.0

# Performance (Phase 9 "Buff")
zstandard>=0.22.0

# OCR (Phase 10 "Lightweight Buff")
rapidocr_onnxruntime>=1.3.0
opencv-python-headless>=4.8.0

# Super Engine (Phase 11 "Data Forge")
polars>=0.20.0
pyarrow>=14.0.0
